requirements.txt
Django==2.2.16
mixer==7.1.2
Pillow==8.3.1
pytest-django==4.4.0
pytest-pythonpath==0.7.3
requests==2.26.0
six==1.16.0
sorl-thumbnail==12.7.0
Faker==12.0.1
Pillow==8.3.1
requests==2.26.0
PyYAML==6.0
flake8==5.0.4
flake8-broken-line==0.5.0
flake8-isort==4.2.0
flake8-plugin-utils==1.3.2
flake8-return==1.1.3
pyTelegramBotAPI==4.10.0
python-dotenv==0.21.1

